scripts/autobuild-run: extend timeout from 120 to 180 minutes
During the last months some package started to timeout, but this is not due
to toolchain infinite loop. I think it's due to the increase of the source
code files' length and to complexity of toolchains or even slow compiling
time. For example python-uvloop and python-lxml only take long time to
build with -Os optimization and this makes it seem a TIMEOUT but it's not.
So let's extend timeout setting from 120 to 180 minutes to avoid false
TIMEOUTs.

Signed-off-by: Giulio Benetti <giulio.benetti@benettiengineering.com>